Feature: current tests

+++ client/components/AllUsers.js
@@ -19,36 +19,41 @@ class AllUsers extends React.Component {
   }
   render() {
     const users = this.props.users
-
-    return (
-      <div className="books-container">
-        {users === undefined || !users.length ? (
-          <h3>Loading...</h3>
-        ) : (
-          users.map(user => (
-            <div className="book-container" key={user.id}>
-              <div className="book-container-left">
-                <p>User's email: {user.email}</p>
-              </div>
-              <div className="book-container-right">
-                {user.isAdmin ? (
-                  <p>This user is an Admin</p>
-                ) : (
-                  <p>This user is not an Admin</p>
-                )}
-                <button
-                  onClick={() => this.handleDelete(user.id)}
-                  type="button"
-                  className="remove_button"
-                >
-                  Remove user
-                </button>
-              </div>
+    if (this.props.user.isAdmin) {
+      return (
+        <div className="books-container">
+          {users === undefined || !users.length ? (
+            <h3>Loading...</h3>
+          ) : (
+            <div>
+              {users.map(user => (
+                <div className="book-container" key={user.id}>
+                  <div className="book-container-left">
+                    <p>User's email: {user.email}</p>
+                  </div>
+                  <div className="book-container-right">
+                    {user.isAdmin ? (
+                      <p>This user is an Admin</p>
+                    ) : (
+                      <p>This user is not an Admin</p>
+                    )}
+                    <button
+                      onClick={() => this.handleDelete(user.id)}
+                      type="button"
+                      className="remove_button"
+                    >
+                      Remove user
+                    </button>
+                  </div>
+                </div>
+              ))}
             </div>
-          ))
-        )}
-      </div>
-    )
+          )}
+        </div>
+      )
+    } else {
+      return <div>You do not have permission!</div>
+    }
   }
 }
